+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## 0.0.1-beta.3
 
+- feat: Add `NewScope` and `NewEvent` functions and use them in the whole codebase
+- feat: Add `AddEventProcessor` to the `Client`
 - ref: Try to read source files from the root directory, based on the filename as well, to make it work on AWS Lambda
 - ref: Remove `gocertifi` dependence and document how to provide your own certificates
+- ref: **[breaking]** Allow for integrations to live on the client, by passing client instance in `SetupOnce` method
+- ref: **[breaking]** Remove `GetIntegration` from the `Hub`
+- ref: **[breaking]** Remove `GlobalEventProcessors` getter from the public API
 
 ## 0.0.1-beta.2